requirements.txt
pyopenxr>=1.1.0
PyOpenGL>=3.1.7
PyOpenGL-accelerate>=3.1.7
Pillow>=10.0.0
numpy>=1.24.0
glfw>=2.6.0
tkinterdnd2>=0.3.0

